Refactor message converter config
 - Move message converter configuration into a separate class `MessageConverterConfigurer`
